ETL_Ncy-main/requirements.txt
pandas>=1.3.0
beautifulsoup4>=4.9.0
requests>=2.25.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.8.0
scikit-learn>=1.0.0
numpy>=1.20.0
python-dotenv>=0.19.0
lxml>=4.6.0
pytest>=6.2.0
loguru>=0.5.0